components/cloud_storage_client_impl/pyproject.toml
[project]
name = "cloud-storage-client-impl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
	"cloud-storage-client-api",
	"google-auth>=2.0.0",
	"google-cloud-storage>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_storage_client_impl"]

[tool.uv.sources]
cloud-storage-client-api = { workspace = true }
